deleted all the check_borders functions, need to start over

+++ libft/ft_substr.c
@@ -12,10 +12,10 @@
 
 #include "libft.h"
 
-char	*ft_substr(char const *s, unsigned int start, size_t len)
+char	*ft_substr(char const *s, int start, int len)
 {
 	char	*subst;
-	size_t	size;
+	int	size;
 
 	if (!s)
 		return (NULL);
