+++ libft/libft.h
@@ -39,7 +39,7 @@ void					*ft_memset(void *b, int c, size_t len);
 char					*ft_strchr(const char *s, int c);
 size_t					ft_strlcat(char *dst, const char *src, size_t size);
 size_t					ft_strlcpy(char *dst, const char *src, size_t size);
-size_t					ft_strlen(const char *c);
+int						ft_strlen(const char *c);
 int						ft_strncmp(const char *s1, const char *s2, size_t n);
 char					*ft_strnstr(const char *big, const char *little,
 							size_t n);
@@ -58,8 +58,8 @@ char					*ft_strmapi(char const *s,
 							char (*f)(unsigned int, char));
 char					*ft_strtrim(char const *s1, char const *set);
 char					*ft_strdup(const char *source);
-char					*ft_substr(char const *s, unsigned int start,
-							size_t len);
+char					*ft_substr(char const *s, int start,
+							int len);
 char					**ft_split(char const *s, char c);
 t_list					*ft_lstnew(void *content);
 void					ft_lstadd_front(t_list **lst, t_list *new);
